* commit '563c3666bb53da98a98befe70d193a378317a8bd':
  Change subId to int from long

+++ src/com/android/mms/data/Conversation.java
@@ -328,7 +328,7 @@ public class Conversation {
                     LogTag.debug("sendReadReport: uri = " + uri);
                 }
                 MmsMessageSender.sendReadRec(context, AddressUtils.getFrom(context, uri),
-                                             c.getString(1), status, c.getLong(2));
+                                             c.getString(1), status, c.getInt(2));
             }
         } finally {
             if (c != null) {
